main_server/pyproject.toml
[tool.poetry]
name = "pyautocontrol"
version = "0.1.0"
description = ""
authors = ["Tasuhuko Kanda"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pyautogui = "^0.9.50"
uvicorn = "^0.11.8"
fastapi = "^0.61.1"
graphene = "^2.1.8"
SQLAlchemy = "^1.3.19"
graphene_sqlalchemy = "^2.3.0"
python-jose = "^3.2.0"
PassLib = "^1.7.2"
python-multipart = "^0.0.5"
aiofiles = "^0.6.0"
PyObjC = { version = "^6.2.2", markers = "sys_platform == 'darwin'" }
PyYAML = "^5.3.1"
websocket-client = "^0.57.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
Image = "^1.5.32"
python3-xlib = "^0.15"
pyscreeze = "^0.1.26"
opencv_python = "^4.3.0.36"
pyautogui = "^0.9.50"
pywin32 = { version = "^228", markers = "sys_platform == 'win32'" }
eel = "^0.13.2"
pytest = "^6.1.2"
opencv-contrib-python = "^4.3.0.36"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
